Reorganization of the code for the AJAX part of the PSA app
