+++ psa/views.py
@@ -5,19 +5,13 @@ from django.contrib.auth.decorators import permission_required
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.views.generic import TemplateView
 from django.utils.translation import gettext as _
-from django.http import JsonResponse
 from django.forms.models import model_to_dict
 from bootstrap_modal_forms.generic import BSModalCreateView, BSModalUpdateView
-from rest_framework.response import Response
-from rest_framework import viewsets, permissions, status
 
 from utils.django.forms import ParaErrorList
-from utils.django.datatables import QueryTableByArgs
 from utils.django.urls import reverse_lazy, http_referer
-from .serializers import CorvetSerializer, CORVET_COLUMN_LIST
 from .forms import NacLicenseForm, NacUpdateIdLicenseForm, NacUpdateForm, CorvetModalForm
 from .models import Corvet, Multimedia
-from .tasks import import_corvet_task
 from dashboard.models import WebLink
 from squalaetp.models import Sivin
 from raspeedi.models import Programing
@@ -115,32 +109,6 @@ class CorvetView(PermissionRequiredMixin, TemplateView):
         return context
 
 
-class CorvetViewSet(viewsets.ModelViewSet):
-    permission_classes = (permissions.IsAuthenticated,)
-    queryset = Corvet.objects.all()
-    serializer_class = CorvetSerializer
-
-    def list(self, request, **kwargs):
-        try:
-            self._filter(request)
-            corvet = QueryTableByArgs(self.queryset, CORVET_COLUMN_LIST, 1, **request.query_params).values()
-            serializer = self.serializer_class(corvet["items"], many=True)
-            data = {
-                "data": serializer.data,
-                "draw": corvet["draw"],
-                "recordsTotal": corvet["total"],
-                "recordsFiltered": corvet["count"]
-            }
-            return Response(data, status=status.HTTP_200_OK)
-        except Exception as err:
-            return Response(err, status=status.HTTP_404_NOT_FOUND)
-
-    def _filter(self, request):
-        query = request.query_params.get('filter', None)
-        if query:
-            self.queryset = Corvet.search(query)
-
-
 @permission_required('psa.view_corvet')
 def corvet_detail(request, pk):
     """
@@ -213,11 +181,3 @@ def product_table(request):
     products = Multimedia.objects.all()
     context.update(locals())
     return render(request, 'psa/product_table.html', context)
-
-
-def import_corvet_async(request):
-    vin = request.GET.get('vin')
-    if vin:
-        task = import_corvet_task.delay(vin=vin)
-        return JsonResponse({"task_id": task.id})
-    return JsonResponse({"nothing to see": "this isn't happening"}, status=400)
